backend/requirements.txt
asgiref==3.8.1
bootstrap4==0.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
defusedxml==0.7.1
Django==5.0.3
django-crispy-forms==2.1
idna==3.7
IMAPClient==3.0.1
oauthlib==3.2.2
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1